webFinanceTracker/requirements.txt
asgiref==3.9.1
cffi==2.0.0
charset-normalizer==3.4.3
contourpy==1.3.3
cryptography==46.0.1
cycler==0.12.1
Django==5.2.6
django-encrypted-model-fields==0.6.5
fonttools==4.60.0
gunicorn==23.0.0
joblib==1.5.2
kiwisolver==1.4.9
matplotlib==3.10.6
numpy==2.3.3
packaging==25.0
pandas==2.3.3
patsy==1.0.1
pillow==11.3.0
pycparser==2.23
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
reportlab==4.4.4
scikit-learn==1.7.2
scipy==1.16.2
six==1.17.0
sqlparse==0.5.3
statsmodels==0.14.5
threadpoolctl==3.6.0
tzdata==2025.2
whitenoise==6.11.0
xgboost==3.0.5
